[Notifications] Enable do not disturb mode when screens are mirrored
When two screens overlap during e.g. a presentation, automatically enable do not disturb mode.
As per usual with automatic do not disturb, it can be overruled by unchecking the checkbox in the notification plasmoid
where it will be disabled until the screen configuration changes again and it still find a mirrored screen setup.

Differential Revision: https://phabricator.kde.org/D22856