Capacity degradation -> Battery health
Summary:
After the change to "Capacity degradation" was announced on
https://pointieststick.com/2019/08/24/kde-usability-productivity-week-85/, it turned out
that nobody could understand the new string either. :(

This patch changes it to something much clearer: suggested by a user: "Battery health".
This string has the advantage that it does not require knowing that batteries degrade,
and uses the same percentage scale as other operating systems.

Test Plan: {F7279584, size=full}

Reviewers: #vdg, #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, aspotashev, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23448