[Kicker/Kickoff] Re-word "Manage <app>..." menu item for greater clarity
Summary:
Even with the Discover icon added in D23623, the text is still fairly unclear,
especially for the common case of wanting to uninstall the app.

This patch re-words the menu item to add that all-important word "Uninstall", while not
losing any of the current meaning.

Test Plan: {F7306465, size=full}

Reviewers: #vdg, #plasma, hein, nicolasfella

Reviewed By: nicolasfella

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23652