[Notifications] Elide application label on the left
Usually it's not very long, so in the common case of just an application name, it won't be truncated anyway.
But in case an origin is set, it might be. In this case, we want to make sure the end of the origin is shown, in case it is for identifying a website domain.

Differential Revision: https://phabricator.kde.org/D23690