[Notifications] Dodge Plasma dialogs
The notification popup typically gets in the way of the system tray popup.
In case it would cover a Plasma dialog, hide the notification popup.
It doesn't specifically check for System Tray but then it will also get out of the way of your
calendar and other widgets you might have placed in the respective corner.

Differential Revision: https://phabricator.kde.org/D24208