Add `sudo` to the lines in install-sessions.sh that need it so you don't need to sudo the whole thing
Summary:
Because this script copies things to /, it needs to be run with `sudo`. But two of the
copy operations also copy things to your home directory, and if you `sudo` the whole
script, this results in those files being copied with root ownership, which can cause
problems later.

This patch prepends `sudo` to just the copy operations that need it so that the ones
that don't aren't run with inappropriately elevated permissions.

Test Plan:
- Run `~/kde/build/plasma-workspace/login-sessions/install-sessions.sh`
- Get prompted for the root password
- Enter it
- See that `plasma-dev-prefix.sh` and `startplasma-dev.sh` in `~/kde/usr/lib64/libexec/` have user ownership, not root ownership

Reviewers: apol, #plasma

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24677