[ksmserver] Drop unused logout effect widget
Summary:
At some point there was an effect in kwin to fade to a grey (with a
weird vigenette effect) as one logs out.

This was signalled with a fake offscreen window with a special window
ID.

The kwin effect has gone, deleted in Plasma 5.8 with
af80a546bfe258b54468b159bb8cf78f16ca9ebf

(Kwin still contains an effect called logout, but that's merely for the
logout prompt)

I intend to replace this logout effect but as I'm going to start from
scratch we may as well communicate from ksmserver -> kwin with a new API
that works on wayland and and solves some other issues all in one.

Test Plan: Compiles

Reviewers: #plasma, #kwin, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24759