Use the same avatar size in lock, login and logout screens
Summary:
Currently the avatar in the logout screen is noticeably smaller than the avatar in the lock and logout screens.

This patch roughly bases the size value on what is produced in the lock and logout, and then applies it to the logout screen.

Test Plan:
Before (logout screen):
{F7680624}

After (logout screen):
{F7680632}

Before (lock screen):
{F7680634}

After (lock screen):
{F7680635}

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: davidedmundson, ngraham, plasma-devel

Tags: #plasma

Maniphest Tasks: T11093

Differential Revision: https://phabricator.kde.org/D25024