[wallpapers/image] Randomise new batches of images in the slideshow
Summary:
There was a concious decision that when we add a new wallpaper we should
append to the existing random order collection, rather than shuffle
existing ones.

The logic is fine but if we start with 0 wallpapers and then add 20 new
ones we don't shuffle anything.

This patch changes it so when a new batch of N wallpapers is added,
those N are shuffled when appending to the existing sort order.

BUG: 413463
FIXED-IN: 5.17.2

Test Plan: qDebug + ran a few times

Reviewers: #plasma, ngraham, davidre

Reviewed By: ngraham, davidre

Subscribers: davidre, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24986