Fix binding loop in lockscreen media controls
Summary:
When we load and position the Image the height is constrained, the width
is not. The image fillMode is set as  PreserveAspectFit so the width
will adjust to fit the aspect ratio,

Therefore we only want to set a fixed height on the sourceSize, it will
automatically match the resultant width.

BUG: 413087
FIXED-IN: 5.17.3

Test Plan:
I was unable to reproduce the state.
Patch is based on reading the code and docs

Reviewers: #plasma, broulik, ngraham

Reviewed By: #plasma, broulik, ngraham

Subscribers: broulik, filipf, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25252