[sddm-theme] Fix initial focus after SDDM QQC2 Port
Summary:
QQC2 StackView has a major behavioural difference

QQuickStackViewPrivate::setCurrentItem
        item->setFocus(true);

This as a general rule makes sense, especially if you're pushing a whole
focus scope onto the stackView, it would automatically select the right
element. However as our base was not a focus scope the
SessionManagementScreen instance itself got focus, taking it away from
the child password box.

LockScreenUi has to be changed so that it sets the initial focus the
same even though the stackview is on QQC1.

Test Plan:
qmlscene in the sddm-theme directory with the dummydata setup
locked the screen regularly

Lots of gammaray

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: davidre, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25261