Keep at least one system monitor alive
Added a ban on disabling the last checkbox in the widget settings.
There will always be at least one graph.

If the user disables the last checkbox, all other checkboxes will automatically get enabled in the settings.
It may be inconvenient for users to uncheck a large number of checkboxes.