battery: Improve the brightness responsiveness
Summary:
Use QQC2.Slider, so that we have a moved signal. This way we can only
issue new brightnesses when the user actually interacts with it.
Don't adapt to the system brightness until we have finished interacting
with it.

Test Plan:
Manual testing, flickering is very much reduced both when scrolling over the
compact plasmoid as well as the slider.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26035