Remove Kirigami DelegateRecycler
Summary:
This is not really needed here due to the size of the delegates and of
the list which is generally short. I ended up getting rid of it because
it makes this config module crash somewhere in the QML runtime when the
user clicks reset. I couldn't track down why DelegateRecycler does this
so far.

Reviewers: #plasma, mart, davidedmundson, crossi, bport

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26241