[applets/digital-clock] Fix truncated Long Date display in a vertical panel
Summary:
BUG: 381551
FIXED-IN: 5.18.0

Test Plan:
Before: {F7851382, size=full}

After: {F7851383, size=full}

Texted extensively because this code is so brittle. No regressions with long date in a
thick or normal-sized horizontal panel or short date in a thick or normal-sized
horizontal or vertical panel.

The one issue I could find is that sometimes switching between short and long date
formats in a double-thickness vertical panel causes the long date  to be truncated again
until plasmashell is restarted. I was unable to determine why. Assistance would be
appreciated if this is a blocker.

Reviewers: #plasma, #vdg, ndavis

Reviewed By: #vdg, ndavis

Subscribers: ndavis, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26312