[Applets/Battery] Don't use toolTipMainText to show info, rather use the second line
Summary:
Currently the applet uses the main line in tooltip to display all info. Instead it should show always the same first line like other widgets and use the second line to show info.

BUG: 411973
FIXED-IN: 5.19.0

Test Plan:
The point of this patch is to make the tooltip more like the style used in the Audio Volume applet:
{F7862465}

Other "after" images:

{F7862461}

{F7862463}

{F7862464}

{F7871067}

Reviewers: #vdg, #plasma, broulik, ndavis, mthw

Reviewed By: #vdg, #plasma, broulik, ndavis, mthw

Subscribers: GB_2, ndavis, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24070