[wallpapers/image] Add default XDG wallpaper locations for slideshow
Summary:
BUG: 415461
FIXED-IN: 5.18.0
Depends on D26510

Right now when you first change the wallpaper plugin to slideshow, no locations are
pre-configured, so you have to add one yourself. This patch adds a
`preferred:wallpaperlocations` token as the default that gets replaced in the user's
config file at runtime with the XDG wallpaper locations.

Test Plan:
Log in as new user > right-click desktop > configure > change to slideshow wallpaper
See that`/usr/share/wallpapers` is already in the list of wallpaper locations; how nice

Reviewers: #plasma, #vdg, ndavis, davidedmundson

Reviewed By: #plasma, #vdg, ndavis, davidedmundson

Subscribers: davidre, broulik, davidedmundson, ndavis, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26483