[wallpapers/image] Add placeholder labels when wallpaper lists are empty
Summary:
Right now, when there are no configured wallpaper slideshow paths or wallpapers in the
slideshow, the list and grid (respectively) are just empty. This patch adds some
placeholder labels to let the user know that it's not broken, just like we do in the
Thunderbolt KCM.

See one `FIXME:` in the code that I could use a hand with.

Test Plan: {F7870947}

Reviewers: #vdg, #plasma, mart

Reviewed By: #vdg, #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26528