[SystemTray] Set maximum size to block tiling
Summary:
System tray dialog/popup can be tiled. This should not be allowed, in this case expanded representaiton should have fixed size. When tiled, it is no longer possible to restore original size.
This also applies to other plasmoids.

BUG: 386551
FIXED-IN: 5.18.0

Test Plan:
1. Open any applet from system tray (for example Clipboard or hidden items)
2. Use shortcut to tile window (by default Meta + Left/Righ/...)
3. Applet will change size and tile at the borded of the screen
4. Expected:
a) before fix: allowed and not possible to restore size
b) after fix: not allowed or restores size on next open

Reviewers: #plasma_workspaces, #plasma, davidedmundson, broulik, ngraham

Reviewed By: ngraham

Subscribers: ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26545