Add global shortcuts for switching to the previous/next activity
Summary:
We currently have a mouse-wheel action to switch to the previous/next activity,
but it wasn't possible to assign a global keyboard shortcut to these actions.

As discussed on Phabricator, this doesn't define a default key-sequence yet,
but at least gives users the possibility to assign there own ones, which wasn't
possible before.

Reviewers: apol, davidedmundson, #vdg, #plasma, ngraham

Reviewed By: davidedmundson, #vdg, #plasma, ngraham

Subscribers: ngraham, romangg, ivan, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D22382