[kcms/feedback] Fix slider value
Summary:
If the kcm's property changes externally we don't update the slider
after the initial load, which is before KConfigModule::load which actually sets
the property.

Whilst there is an alternative fix calling load() a second time, we need
this anyway for default and reset buttons to work correctly.

I also think the original code was broken, I don't see why we would map
the value back to modeOptions[idx], we want the slider value to be the
index of the array, not the enum value. They just happen to be the same.

Test Plan: Opened KCM changed it, opened KCM repeat

Reviewers: #plasma, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26678