[sddm-theme] Don't have a broken reveal password button
Summary:
sddm-greeter will have a button for the reveal password button, but due
to sddm-greeter not loading a relevant QPT has no code to force it to
load the breeze icon set.

Without the breeze icon set, the clear button does not show.

There are ways to solve this, but none are trivial or reliable.

I threatened to do a revert in 5.12 (https://phabricator.kde.org/D9040)
but the bug has still not been fixed since.

BUG: 396039
CCBUG: 387418

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26675