[context menu] Remove last traces of "Lock Widgets" action from UI
Summary:
In Plasma, 5.18, this ishn't a user-facing option anymore. It's still visible in the
context meny configuration dialog though, and it's still referenced in some of the
context menu building code.

This patch removes it from those places

BUG: 416578
FIXED-IN: 5.18.0

Test Plan:
Context menu still looks the same on the desktop and panel, both in and out of edit mode
The action no longer shows up in the configuration UI.

Please review carefully as this code is fairly unfamiliar to me.

Reviewers: mart, #plasma

Reviewed By: mart, #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D26845