Block calls to org.kde.KSplash till it is registered
Summary:
A recent update to startup made it too fast and hit a race with ksplash.

We expose that we could send that kinit and ksmserver were ready before
ksplash registered the service name.

We have an existing mechanism to make DBus server queue up messages if a
service is not registered yet without actually activating it. So we can
re-use that.

Long term we need to rethink ksplash completely.

Test Plan:
Confirmed issue with extra debug
Logged in and the issue was gone

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27702