Fixed margin of digital clock
Summary: The clock was too close to the panel right edge when displaying hours beginning with 10, 11 and 12. Weirdly enough, the right margin was different based on the clock current time. This patch makes the margin constantly equal to the one that there was before when displaying 10,11,12 plus units.smallSpacing.

Reviewers: #vdg, #plasma, davidedmundson, ngraham

Reviewed By: #vdg, #plasma, davidedmundson, ngraham

Subscribers: gvgeo, davidedmundson, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27737