[plasma-session] Continue to start kdeinit
Summary:
This was a deliberate decision to remove this in
c673a832263036e239587c9a2bbe57e2209db905

The logic being that kdeinit should auto launch itself if required.

In practice this exposes a presumably pre-existing bug in kdeinit
autostarting where the env is not set correctly.

Test Plan:
 qdbus org.kde.klauncher5 /KLauncher org.kde.KLauncher.exec_blind /opt/kde5/bin/kate ""

kate opened

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27868