[gmenu-dbusmenu-proxy] Pass action "target" in invocation
According to documentation [1] a menu item has a target attribute, which is "the parameter to pass when activating the action".
Furthermore, the action "Activate" method specifies "If the action activation requires a parameter then this parameter must
be given in the second parameter (av).". Also implements a TODO in the code.

[1] https://wiki.gnome.org/Projects/GLib/GApplication/DBusAPI

CCBUG: 418385

Differential Revision: https://phabricator.kde.org/D27884