[gmenu-dbusmenu-proxy] Fix radio button state detection
Instead of comparing the state of the action to the action name, compare it to the argument the action will invoke it with.
In contrast to checkboxes, which are independent actions with a boolean state, radio menu items all refer
to the same action (e.g. "Zoom") and then pass the new state ("zoom-100", "zoom-200", etc) as target when invoking the action.

CCBUG: 418385

Differential Revision: https://phabricator.kde.org/D27885