[kioslave/desktop] Attempt at making test more robust
Summary:
We have an unstable test.
KCoreDirLister has a timer involved set to 200ms

QTRY_VERIFY will fire when we have one change from the spy, but maybe
not everything if we explicitly wait and process maybe it will improve
things?

It's a bit of a random shot-in the dark.

Test Plan: Passes locally..but it always passed locally before

Reviewers: #plasma, kmaterka

Reviewed By: kmaterka

Subscribers: kmaterka, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28183