Remove unnecessary code and function calls
Summary:
The ShellConfig class has been deleted. This config class was in KDE4 for the runOptions widget used and the class/the settings are not used anymore.
Additionally the context type check has been removed, because this is already solved using the ignore types.
Futhermore the match icon and the list of actions can be reused.

Test Plan: Should compile and execute shell commands as before.

Reviewers: davidedmundson, ngraham, broulik, meven, #plasma

Reviewed By: meven

Subscribers: meven, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27495