Make Pin buttons consistently sized in System Tray and Clock pop-ups
Summary:
The Pin buttons in the System Tray and Clock pop-ups are very small. This has
always felt slightly odd to me, and the recent change to add a defined header
area for the System Tray pop-up makes it even odder in my opinion, as the icon
is now much smaller than the area it visibly inhabits. This patch makes both icons
in both applets use the standard size with no local overrides. The clock pop-up's
pin button is ported to PC3 in the process.

Depends on D28228

Test Plan:
Before: {F8194101}
After: {F8194100}

Reviewers: #vdg, #plasma, ognarb, mart

Reviewed By: #vdg, #plasma, ognarb, mart

Subscribers: mart, ndavis, broulik, ognarb, plasma-devel

Tags: #plasma

Maniphest Tasks: T10470

Differential Revision: https://phabricator.kde.org/D28211