[applets/devicenotifier] Port to ExpandableListItem
Summary:
Port the Device Notifier to use an ExpandibleListItem.

Tested with:
- Removable flash drive
- Removable LUKS-encrypted hard disk drive
- Android phone
- iOS phone
- Internal disks

No regressions or departures from the current behavior were found.

The only major UI change is the loss of the progress bar, which is difficult
to display properly with an ExpandableListItem. It is replaced with a textual
display of free space and total space in the subtitle, which is more consistent with the
appearance of other ExpandableListItem-using applets, and IMO is more useful anyway.