[applets/digital-clock] Fix calendar events not being shown at first
Summary:
I noticed that on the first time I bring up the calendar by clicking the clock on my Plasma panel, it always shows "No event for today". If I click another day on the calendar and then click back on today, it shows today's events fine. It should display it correctly the first time.

To fix this bug, we need to compare the date without time.

BUG: 395956
FIXED-IN: 5.19.0

Reviewers: ngraham, #plasma, vkrause

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28354