[applets/clipboard]: Fix applet closing itself when clicking on a history item while the tray popup is pinned open
Summary:
When I pin a Klipper widget and then click on a history item, the Klipper widget closes.

To fix this bug, we need to check if Klipped is pinned, but to be able to do this, we also need to "forward" the pinned state from the SystemTray applet to the Klipper applet.

BUG: 416510
FIXED-IN: 5.19.0

Reviewers: ngraham, #plasma

Reviewed By: ngraham

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28408