Make headings of systray and applet merge
Summary:
When the applet is a page with a plasmoidHeading as header, merge it with systray one.
{F8215920}

The only big problem is that it looks bad when you open said applet while in sidebar view, see:
{F8215916}

I'd like to solve that by removing the sidebar and adding a back button, but that would be for another patch in the future.

Test Plan: To be tested together with D28467

Reviewers: #plasma, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: mart, manueljlin, ognarb, ngraham, plasma-devel

Tags: #plasma

Maniphest Tasks: T10470

Differential Revision: https://phabricator.kde.org/D28575